refactor(meal): move MealPlanScreen to impl, delegate via BlocScreen
Same recipe. MealPlanBloc extends BlocScreen; MealPlanBlocImpl
overrides Content(). BottomNavBloc.Child.Meals delegates.

meal/core/public sets publicResClass = true; meal/core/impl
gains the Compose plugin + ui-deps.
