+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavScreen.kt
@@ -42,7 +42,6 @@ import com.arkivanov.decompose.extensions.compose.stack.animation.predictiveback
 import com.arkivanov.decompose.extensions.compose.stack.animation.scale
 import com.arkivanov.decompose.extensions.compose.stack.animation.stackAnimation
 import com.plusmobileapps.chefmate.browser.BrowserRootScreen
-import com.plusmobileapps.chefmate.meal.core.MealPlanScreen
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.BROWSER
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.GROCERIES
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc.Tab.MEALS
@@ -152,7 +151,7 @@ private fun BottomNavContentContainer(bloc: BottomNavBloc, modifier: Modifier = 
         when (val instance = created.instance) {
             is BottomNavBloc.Child.Browser -> BrowserRootScreen(instance.bloc)
             is BottomNavBloc.Child.GroceryList -> instance.Content()
-            is BottomNavBloc.Child.Meals -> MealPlanScreen(instance.bloc)
+            is BottomNavBloc.Child.Meals -> instance.Content()
             is BottomNavBloc.Child.RecipeList -> instance.Content()
             is BottomNavBloc.Child.Settings -> instance.Content()
         }
