+++ client/meal/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/meal/core/MealPlanBloc.kt
@@ -6,10 +6,11 @@ import com.plusmobileapps.chefmate.meal.data.MealPlanItem
 import com.plusmobileapps.chefmate.recipe.core.addmeal.MealPlannerRootBloc
 import com.plusmobileapps.chefmate.text.FixedString
 import com.plusmobileapps.chefmate.text.TextData
+import com.plusmobileapps.chefmate.ui.BlocScreen
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.datetime.LocalDate
 
-interface MealPlanBloc {
+interface MealPlanBloc : BlocScreen {
     val state: StateFlow<Model>
 
     fun onViewModeSelected(mode: ViewMode)
